requirements.txt
flask
flask-cors
tensorflow
Pillow
numpy
mysql-connector-python
gdown
